fixed a heal typo.
change the actual HiH controling routines, now the checkbox 'disable heal if health is below 20' overrides the other 3 options, so you can stay doa for a while.
heal after 5 min is over ridden by the above box, but when its timer is 0 will heal regardless of lower 2.
timer is set from being attacked ONLY.
if the script fights till you die you will still heal without waiting.
heal if stamina is full or can be spent, one or both have to be set to heal with the top option to disable if below 20